+++ plugin-hrm-form/src/HrmFormPlugin.tsx
@@ -47,7 +47,7 @@ import { setupConferenceComponents, setUpConferenceActions } from './conference'
 import { setUpTransferActions } from './transfer/setUpTransferActions';
 import { playNotification } from './notifications/playNotification';
 import { namespace } from './states/storeNamespaces';
-import { setUpMaskedIdentifiers } from './components/common/MaskingIdentifiers';
+// import { setUpMaskedIdentifiers } from './components/common/MaskingIdentifiers';
 
 const PLUGIN_NAME = 'HrmFormPlugin';
 
@@ -78,9 +78,26 @@ const setUpComponents = (
   translateUI: (language: string) => Promise<void>,
 ) => {
   const { canView } = getPermissionsForViewingIdentifiers();
-  const maskIdentifiers = !canView(PermissionActions.VIEW_IDENTIFIERS);
+  const maskIdentifiers = true // !canView(PermissionActions.VIEW_IDENTIFIERS);
 
-  if (maskIdentifiers) setUpMaskedIdentifiers();
+  // if (maskIdentifiers) setUpMaskedIdentifiers();
+  if(maskIdentifiers) {
+    Channels.maskIdentifiersForDefaultChannels();
+
+  // Mask the username within the messable bubbles in an conversation
+  Flex.MessagingCanvas.defaultProps.memberDisplayOptions = {
+    theirDefaultName: 'XXXXXX',
+    theirFriendlyNameOverride: false,
+    yourFriendlyNameOverride: true,
+  };
+  Flex.MessageList.Content.remove('0');
+  // Masks TaskInfoPanelContent - TODO: refactor to use a react component
+  const strings = getTemplateStrings();
+  strings.TaskInfoPanelContent = strings.TaskInfoPanelContentMasked;
+  strings.CallParticipantCustomerName = strings.MaskIdentifiers;
+
+  Components.setUpViewMaskedVoiceNumber();
+  }
 
   // setUp (add) dynamic components
   Components.setUpQueuesStatusWriter(setupObject);
